Remove redundant button maps
This change removes redundant profiles from the XML button maps
in preparation for the codex approach where a single button map
defines mappings for all controllers.

Effort was made to preserve the correct profile in the buttonmap,
e.g. a SNES controller with button maps for Default and SNES will
keep the SNES profile and drop the Default profile.